Implemented a fbEmbed function (read desc)
Ok so, H U G E change. Not featurewise, but codewise. I realized that in embeds, i  did `setColor()` like 20 thousand times + copying the tick and cross emojis whenever i had to use them

My initial idea to kinda solve this was to just store the mojis in a variable, but i decided to make an entire function which takes in the type, title, and desc. The type determines the colour of the embed and the emoji that the title starts of with (like the tick emoji etc)

for now i have 2 types which are success and error, but could add more later in the future.